requirements.txt
aiocache==0.2.2
aiofiles==0.3.0
aiomcache==0.4.0
aiosmtpd==1.0a5
atpublic==0.5
cassandra-driver==3.7.1
docutils==0.13.1
httptools==0.0.9
jmespath==0.9.0
multidict==2.1.4
python-dateutil==2.6.0
requests==2.12.4
Sanic==0.1.8
six==1.10.0
ujson==1.35
uvloop==0.6.7
validate-email==1.3
arrow==0.10.0
